# api/openapi.yaml
openapi: 3.1.0
info:
  title: NATS Consol API
  version: 0.8.0
  description: Self-hosted admin console for NATS JetStream — cluster-scoped REST, WebSocket, and SSE API.
servers:
  - url: /
    description: API root
security:
  - basicAuth: []
  - bearerAuth: []
  - sessionCookie: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login with username and password
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: { type: string }
                password: { type: string, format: password }
      responses:
        "200":
          description: Logged in; sets session, refresh, and CSRF cookies
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Rotate access and refresh tokens
      description: Requires the refresh cookie and matching CSRF token. Validates the device fingerprint (User-Agent + client IP).
      security: []
      responses:
        "200":
          description: New session and refresh cookies issued
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: List alerts in the feed
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [open, closed] }
        - name: clusterId
          in: query
          schema: { type: string, format: uuid }
        - name: severity
          in: query
          schema: { type: string, enum: [info, warning, critical] }
        - name: limit
          in: query
          schema: { type: integer }
      responses:
        "200":
          description: Alert list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertListResponse"
...
  /api/v1/alerts/{alertId}:
    get:
      tags: [Alerts]
      summary: Get alert detail
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertResponse"
  /api/v1/alerts/{alertId}/acknowledge:
    post:
      tags: [Alerts]
      summary: Acknowledge an open alert (hides from badge)
      parameters:
        - $ref: "#/components/parameters/AlertId"
      responses:
        "200":
          description: Acknowledged alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertResponse"
...
  /api/v1/alert-rules/{ruleId}:
    get:
      tags: [Alerts]
      summary: Get alert rule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Rule
    patch:
      tags: [Alerts]
      summary: Update alert rule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "200":
          description: Updated rule
    delete:
      tags: [Alerts]
      summary: Delete alert rule
      parameters:
        - $ref: "#/components/parameters/RuleId"
      responses:
        "204":
          description: Deleted
  /api/v1/audit:
    get:
      tags: [Audit]
      summary: List audit log entries (admin)
      parameters:
        - name: cluster_id
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Audit entries
...
  /api/v1/users/{userId}/roles:
    put:
      tags: [Users]
      summary: Set user roles (admin)
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Updated user
...
  /api/v1/clusters/{clusterId}:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Clusters]
      summary: Get cluster
      responses:
        "200":
          description: Cluster details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clusters/{clusterId}/test:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    post:
      tags: [Clusters]
      summary: Test cluster connectivity
      responses:
        "200":
          description: Connection test result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClusterTestResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clusters/{clusterId}/account:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [JetStream]
      summary: JetStream account info
      responses:
        "200":
          description: Account info
          content:
            application/json:
              schema:
                type: object
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/event-catalog:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [EventCatalog]
      summary: List event catalog
      description: >
        Hybrid Swagger-style catalog of concrete NATS subjects discovered from
        JetStream stream subjects and consumer filters, merged with Postgres
        enrichments (owner, description, JSON Schema, example, deprecation).
        Consumers whose filters match each subject are included. Use `?fresh=1`
        to bypass the monitoring cache.
      parameters:
        - name: fresh
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Bypass cached jsz snapshot when set to 1
      responses:
        "200":
          description: Event catalog snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCatalogSnapshot"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/event-wikipedia:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [EventWikipedia]
      summary: List event Wikipedia pages
      description: >
        Auto-assembled per-subject documentation pages built from Event Catalog
        enrichments, live JetStream consumers, and Event Genome related peers.
        Use `?fresh=1` to bypass the monitoring cache and `?subject=` to filter
        a single page.
      parameters:
        - name: fresh
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Bypass cached jsz snapshot when set to 1
        - name: subject
          in: query
          schema:
            type: string
          description: Optional concrete subject to return a single Wikipedia page
      responses:
        "200":
          description: Event Wikipedia snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventWikipediaSnapshot"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/event-catalog/{subject}:
    parameters:
      - $ref: "#/components/parameters/clusterId"
      - name: subject
        in: path
        required: true
        schema:
          type: string
        description: Concrete event subject (no NATS wildcards)
    put:
      tags: [EventCatalog]
      summary: Upsert event catalog enrichment
      description: Create or update owner, description, JSON Schema, example, and deprecation for a subject.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCatalogUpsert"
      responses:
        "200":
          description: Saved enrichment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCatalogDoc"
        "400":
          $ref: "#/components/responses/BadRequest"
    delete:
      tags: [EventCatalog]
      summary: Delete event catalog enrichment
      description: Removes documentation only; live subjects still appear as undocumented.
      responses:
        "204":
          description: Enrichment removed
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/clusters/{clusterId}/hidden-bottlenecks:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Monitoring]
      summary: Find hidden bottlenecks (schedule × payload correlations)
      description: >
        Deterministic mining of recurring weekday×hour patterns from hourly
        rollups (consumer lag, derived avg payload, optional fingerprint
        processing latency). Not the same as slow-consumer lag/CPU threshold
        alerts. Pass demo=1 for the canned Friday 18:00 showcase.
      parameters:
        - name: demo
          in: query
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: Bottleneck snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HiddenBottleneckSnapshot"
  /api/v1/clusters/{clusterId}/hidden-bottlenecks/ask:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    post:
      tags: [Monitoring]
      summary: Ask AI to narrate hidden bottleneck findings
      description: >
        Gemini narrates the same precomputed snapshot as GET (does not invent findings).
      parameters:
        - name: demo
          in: query
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                message:
                  type: string
      responses:
        "200":
          description: AI reply plus snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  reply:
                    type: string
                  snapshot:
                    $ref: "#/components/schemas/HiddenBottleneckSnapshot"
...
  /api/v1/clusters/{clusterId}/chaos-story:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Docs]
      summary: Chaos story seed or demo
      description: >
        Without demo=1, returns capped live inventory names for the Chaos Story
        Generator. With demo=1, returns the canned Black Friday story.
      parameters:
        - name: demo
          in: query
          schema:
            type: string
            enum: ["1"]
        - name: fresh
          in: query
          schema:
            type: string
            enum: ["1"]
      responses:
        "200":
          description: Story and/or seed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChaosStoryEnvelope"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/chaos-story/generate:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    post:
      tags: [Docs]
      summary: Generate a realistic chaos story with AI
      description: >
        Gemini invents a multi-act disaster narrative using only stream/consumer/subject
        names from live inventory. Does not inject faults. Requires AI_ENABLED.
      parameters:
        - name: fresh
          in: query
          schema:
            type: string
            enum: ["1"]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                hint:
                  type: string
                  description: Optional operator hint (e.g. Black Friday)
      responses:
        "200":
          description: Generated story plus seed
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: object
                    properties:
                      story:
                        $ref: "#/components/schemas/ChaosStory"
                      seed:
                        $ref: "#/components/schemas/ChaosStorySeed"
        "404":
          description: AI assistant not enabled
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/incident-annotations:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    post:
      tags: [Monitoring]
      summary: Create an incident annotation (e.g. deploy marker)
      description: >
        Records a CI/manual change event used by incident reconstruction.
        When deploy annotations exist in a window they take precedence over
        audit-log change fallbacks.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/IncidentAnnotationCreate"
      responses:
        "201":
          description: Annotation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IncidentAnnotation"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/v1/clusters/{clusterId}/monitoring/varz:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Monitoring]
      summary: NATS varz
      responses:
        "200":
          description: varz JSON
          content:
            application/json:
              schema:
                type: object
  /api/v1/clusters/{clusterId}/replicas:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Monitoring]
      summary: NATS server replicas (view-only)
      description: |
        Projects cluster server peers from varz, routez, and jsz meta_cluster.
        Rich varz stats apply only to the monitored node; other peers are route/meta-derived.
      parameters:
        - name: fresh
          in: query
          schema:
            type: string
            enum: ["1"]
          description: Bypass snapshot hub cache for varz/jsz
      responses:
        "200":
          description: Replica snapshot
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/ReplicasSnapshot"
        "502":
          $ref: "#/components/responses/BadGateway"
  /api/v1/clusters/{clusterId}/replicas/events:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Monitoring]
      summary: Replicas SSE (demand-driven)
      description: |
        Server-Sent Events stream of projected replica snapshots while the Replicas
        page is open. Scrapes about every 2s while at least one subscriber is connected.
      responses:
        "200":
          description: text/event-stream with event name "replicas"
          content:
            text/event-stream:
              schema:
                type: string
        "503":
          description: replicas broker unavailable
  /api/v1/clusters/{clusterId}/monitoring/jsz:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Monitoring]
      summary: NATS jsz
      parameters:
        - name: acc
          in: query
          schema:
            type: string
        - name: streams
          in: query
          schema:
            type: boolean
        - name: consumers
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: jsz JSON
          content:
            application/json:
              schema:
                type: object
  /api/v1/clusters/{clusterId}/streams:
    parameters:
      - $ref: "#/components/parameters/clusterId"
    get:
      tags: [Streams]
      summary: List streams
      responses:
        "200":
          description: Stream list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StreamListResponse"
    post:
      tags: [Streams]
      summary: Create stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamConfig"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
  /api/v1/clusters/{clusterId}/streams/{name}:
    parameters:
      - $ref: "#/components/parameters/clusterId"
      - $ref: "#/components/parameters/streamName"
    get:
      tags: [Streams]
      summary: Get stream
      responses:
        "200":
          description: Stream info
          content:
            application/json:
              schema:
                type: object
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Streams]
      summary: Update stream
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamConfig"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                type: object
    delete:
      tags: [Streams]
      summary: Delete stream
      responses:
        "204":
          description: Deleted
  /api/v1/clusters/{clusterId}/streams/{name}/impact:
    parameters:
      - $ref: "#/components/parameters/clusterId"
      - $ref: "#/components/parameters/streamName"
    get:
      tags: [Streams]
      summary: Blast radius analysis for deleting a stream
      description: >
        Returns impacted service counts, related streams (mirror/source dependents
        and companion DLQ), consumer count, and critical service names derived from
        consumer metadata (nats-consol/service, nats-consol/critical).
      responses:
        "200":
          description: Blast radius summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlastRadius"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/clusters/{clusterId}/streams/{name}/purge:
    parameters:
      - $ref: "#/components/parameters/clusterId"
      - $ref: "#/components/parameters/streamName"
    post:
      tags: [Streams]
      summary: Purge stream
      responses:
        "204":
          description: Purged
...
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: RS256 session JWT from POST /api/v1/auth/login (same claims as the session cookie).
    sessionCookie:
      type: apiKey
      in: cookie
      name: nats_consol_session
      description: HttpOnly RS256 session JWT cookie set by login.
  parameters:
    clusterId:
      name: clusterId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    AlertId:
      name: alertId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    RuleId:
      name: ruleId
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadGateway:
      description: Upstream NATS error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    AlertListResponse:
      type: object
      properties:
        data:
          type: array
          items: { $ref: "#/components/schemas/Alert" }
        meta:
          $ref: "#/components/schemas/Meta"
...
    AlertResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Alert"
...
    ClusterResponse:
      type: object
      properties:
        data:
          $ref: "#/components/schemas/Cluster"
...